feat: hoist submissions log to team level (#4305)
