+++ editor.planx.uk/src/pages/layout/AuthenticatedLayout.tsx
@@ -1,7 +1,7 @@
 import Box from "@mui/material/Box";
 import { containerClasses } from "@mui/material/Container";
 import { styled } from "@mui/material/styles";
-import EditorNavMenu from "components/EditorNavMenu";
+import EditorNavMenu from "components/EditorNavMenu/EditorNavMenu";
 import RouteLoadingIndicator from "components/RouteLoadingIndicator";
 import { useStore } from "pages/FlowEditor/lib/store";
 import React, { PropsWithChildren } from "react";
