+++ editor.planx.uk/src/routes/submissionsLog.tsx
@@ -1,22 +1,25 @@
 import { compose, mount, route, withData } from "navi";
 import Submissions from "pages/FlowEditor/components/Settings/Submissions";
+import React from "react";
 
 import { makeTitle } from "./utils";
 
 const submissionsLogRoutes = compose(
   withData((req) => ({
     mountpath: req.mountpath,
-    flow: req.params.flow.split(",")[0],
+    flow: req.params.flow?.split(",")[0],
   })),
 
   mount({
     "/": compose(
-      route(async (req) => ({
-        title: makeTitle(
-          [req.params.team, req.params.flow, "submissions-log"].join("/"),
-        ),
-        view: Submissions,
-      })),
+      route(async (req) => {
+        const { team: teamSlug, flow: flowSlug } = req.params;
+
+        return {
+          title: makeTitle([teamSlug, "submissions-log", flowSlug].join("/")),
+          view: <Submissions flowSlug={flowSlug} />,
+        };
+      }),
     ),
   }),
 );
