+++ editor.planx.uk/src/routes/team.tsx
@@ -90,9 +90,7 @@ const routes = compose(
 
     "/:flow/service": setFlowAndLazyLoad(() => import("./serviceSettings")),
 
-    "/:flow/submissions-log": setFlowAndLazyLoad(
-      () => import("./submissionsLog"),
-    ),
+    "/submissions-log": lazy(() => import("./submissionsLog")),
 
     "/members": lazy(() => import("./teamMembers")),
     "/design": compose(
