+++ hasura.planx.uk/metadata/tables.yaml
@@ -1893,15 +1893,16 @@
     - role: demoUser
       permission:
         columns:
-          - retry
-          - team_id
-          - response
+          - created_at
           - event_id
           - event_type
-          - status
-          - created_at
           - flow_id
+          - flow_name
+          - response
+          - retry
           - session_id
+          - status
+          - team_id
         filter:
           team:
             flows:
@@ -1911,29 +1912,31 @@
     - role: platformAdmin
       permission:
         columns:
-          - retry
-          - team_id
-          - response
+          - created_at
           - event_id
           - event_type
-          - status
-          - created_at
           - flow_id
+          - flow_name
+          - response
+          - retry
           - session_id
+          - status
+          - team_id
         filter: {}
       comment: ""
     - role: teamEditor
       permission:
         columns:
-          - retry
-          - team_id
-          - response
+          - created_at
           - event_id
           - event_type
-          - status
-          - created_at
           - flow_id
+          - flow_name
+          - response
+          - retry
           - session_id
+          - status
+          - team_id
         filter:
           team:
             members:
